Create ultrasonic sensor
In this example, we use pins PB0 and PB1 for the trigger and echo signals of the ultrasonic sensor, respectively. The DDR registers are used to set the direction of these pins as output and input, respectively.

Timer 1 is used to measure the pulse width of the echo signal. The input capture interrupt (TIMER1_CAPT_vect) is used to capture the value of the timer when the echo signal transitions from low to high. The pulse width is then calculated by subtracting the captured value from the current value of the timer.

Finally, the distance in centimeters is calculated by dividing the pulse width by the speed of sound (approximately 340 m/s) and then dividing by 2 (since the sound travels twice the distance from the sensor to the object and back).

Note that the exact values used for the trigger pulse width and the timer presales depend on the specific ultrasonic sensor being used. You may need to adjust these values to get accurate distance measurements.